+++ packages/components/src/Workflow/subcomponents/Main/Main.tsx
@@ -1,6 +1,7 @@
 import React, { HTMLAttributes, ReactNode } from "react"
 import classnames from "classnames"
 import { OverrideClassName } from "~types/OverrideClassName"
+import styles from "./Main.module.scss"
 
 export type WorkflowMainProps = OverrideClassName<
   HTMLAttributes<HTMLDivElement>
@@ -13,7 +14,7 @@ export const Main = ({
   classNameOverride,
   ...restProps
 }: WorkflowMainProps): JSX.Element => (
-  <main className={classnames([classNameOverride])} {...restProps}>
+  <main className={classnames([styles.main, classNameOverride])} {...restProps}>
     {children}
   </main>
 )