+++ packages/components/src/index.ts
@@ -1,5 +1,14 @@
+export * from "./ClearButton"
 export * from "./Filter"
+export * from "./FilterBar"
 export * from "./FilterButton"
+export * from "./FilterDatePicker"
+export * from "./FilterDateRangePicker"
+export * from "./FilterMultiSelect"
 export * from "./FilterSelect"
+export * from "./InputSearch"
 export * from "./KaizenProvider"
 export * from "./Workflow"
+
+// TODO: Remove when Date Picker is migrated to KAIO
+export * from "./types/DatePicker"