+++ storybook/components/index.ts
@@ -1,3 +1,4 @@
+export * from "./NoClipCanvas"
 export * from "./DocsContainer/ComponentDocsTemplate/subcomponents/KaioInstallation/KaioNotification"
 export * from "./Installation"
 export * from "./ResourceLinks"